Adds support for correlated subqueries
Correlated subqueries [1] were prevented by validation which asserted
that the field filtered existed in the `_from` tables. There was a work
around which was to temporarily add the table to the _from list, then
create the correlated subquery, then later remove it.

This approach introduces a way to specifically highlight that the query
being built is a correlated subquery, and as such include the outer
table at the time of creation with `from_`.

An alternate approach may have been to introduce a flag which disabled
table based validation on fields when creating new WHERE clauses. This
approach was decided against because a correlated subquery requires
namespaces lest there be a clash when filtering "id" against "id" in the
outer table.