Removes exception raising validate_term
In order to support correlated subqueries, we no longer raise an exception
when a filter clause references a table not included in the FROM or JOIN
fields. Instead a new function _validate_table is born which returns
True if the term references a FROM or JOIN clause table, and False
otherwise.

We then use the full namespace when dealing with this query, because the
reference to parent table is likely to cause an ambiguous scenario.