Allow raw sql string usage
As some dialects don't understand the current functions, it's a way to
use some reserved functions in all dialects without the need to change
the core for each.

I understand that with it we can open a margin for errors, but also we
can give to the user more flexibility.

Signed-off-by: Airton Zanon <me@airton.dev>