Implement CustomFunction
As some dialects don't understand the current functions, it's a way to
use some reserved functions in all dialects without the need to change
the core for each.

This works as a Factory, it gets the function name and arguments and
return a callable that can be used in the query builder.

It also works for functions that don't need arguments.

Signed-off-by: Airton Zanon <me@airton.dev>