Support dialects in Table
This commit allows the `Table` class to be initialized with an optional
`query_cls` parameter, which is used in place of the standard `Query`
class by the `select`, `update` and `insert` methods.

This commit also adds a couple of new convenience methods on the `Query`
class to allow easier creation of tables that use a specific `Query`
subclass.

Signed-off-by: Petter Nyström <jimorie@gmail.com>