make insert_dict() api less confusing
in SQL builders, order of calling individual functions should
not matter. Originally it was allowed to first call columns()
and then insert_dict as long as the two matched, but I'm not even sure
it worked, as we should have used set comparison instead.

It is also better to have symmetric API in