Improve typing for builder methods (#850)
* Improve typing for builder

Add overloads and typing for builder parameters to help type checkers
and improve completions for chained methods.

E.g. functions.Count('*').distinct() return type would be inferred
as None, disabling completions for further chaining and rendering an
error when calling .as_(...) on it

* Make joiner generic to keep the original query builder type after on clause

* Replace explicit return types of builders with None

Updated all builder-decorated methods' signatures that return the same copied object (effectively self) with None return type, so that the correct return type is determined by the decorator.

This eliminates lots of typecheck errors and more importantly allows to get the correct typing of builder method in overridden classes